// web/src/generated/models/agentDefinitionCreated.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Priv-View Core API
 * Phase-one Core API for authentication, hierarchical tenants, clients, matters, and matter metadata definitions.
 * OpenAPI spec version: 0.1.0
 */
import type { AgentDefinitionRead } from './agentDefinitionRead';
import type { AgentDefinitionVersionRead } from './agentDefinitionVersionRead';

export interface AgentDefinitionCreated {
  agent: AgentDefinitionRead;
  version: AgentDefinitionVersionRead;
}