// web/src/generated/models/agentDefinitionUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Priv-View Core API
 * Phase-one Core API for authentication, hierarchical tenants, clients, matters, and matter metadata definitions.
 * OpenAPI spec version: 0.1.0
 */
import type { AgentDefinitionUpdateStatus } from './agentDefinitionUpdateStatus';

export interface AgentDefinitionUpdate {
  name?: string | null;
  description?: string | null;
  status?: AgentDefinitionUpdateStatus;
}